move public utils to common file
