+++ test/e2e/staged_updaterun_test.go
@@ -17,7 +17,6 @@ limitations under the License.
 package e2e
 
 import (
-	"context"
 	"fmt"
 	"time"
 
@@ -25,7 +24,6 @@ import (
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
@@ -1671,48 +1669,4 @@ func createStagedUpdateRunSucceedWithNoResourceSnapshotIndex(updateRunName, name
 		},
 	}
 	Expect(hubClient.Create(ctx, updateRun)).To(Succeed(), "Failed to create StagedUpdateRun %s", updateRunName)
-}
-
-func UpdateStagedUpdateRunState(ctx context.Context, hubClient client.Client, updateRunName, namespace string, state placementv1beta1.State) {
-	Eventually(func() error {
-		updateRun := &placementv1beta1.StagedUpdateRun{}
-		if err := hubClient.Get(ctx, types.NamespacedName{Name: updateRunName, Namespace: namespace}, updateRun); err != nil {
-			return fmt.Errorf("failed to get StagedUpdateRun %s", updateRunName)
-		}
-
-		updateRun.Spec.State = state
-		if err := hubClient.Update(ctx, updateRun); err != nil {
-			return fmt.Errorf("failed to update StagedUpdateRun %s", updateRunName)
-		}
-		return nil
-	}, eventuallyDuration, eventuallyInterval).Should(Succeed(), "Failed to update StagedUpdateRun %s to state %s", updateRunName, state)
-}
-
-func ValidateAndApproveNamespacedApprovalRequests(ctx context.Context, hubClient client.Client, updateRunName, namespace, stageName, approvalRequestNameFmt, stageTaskType string) {
-	Eventually(func() error {
-		appReqList := &placementv1beta1.ApprovalRequestList{}
-		if err := hubClient.List(ctx, appReqList, client.InNamespace(namespace), client.MatchingLabels{
-			placementv1beta1.TargetUpdatingStageNameLabel: stageName,
-			placementv1beta1.TargetUpdateRunLabel:         updateRunName,
-			placementv1beta1.TaskTypeLabel:                stageTaskType,
-		}); err != nil {
-			return fmt.Errorf("failed to list approval requests: %w", err)
-		}
-
-		if len(appReqList.Items) != 1 {
-			return fmt.Errorf("got %d approval requests, want 1", len(appReqList.Items))
-		}
-		appReq := &appReqList.Items[0]
-		approvalRequestName := fmt.Sprintf(approvalRequestNameFmt, updateRunName, stageName)
-		if appReq.Name != approvalRequestName {
-			return fmt.Errorf("got approval request %s, want %s", appReq.Name, approvalRequestName)
-		}
-		meta.SetStatusCondition(&appReq.Status.Conditions, metav1.Condition{
-			Status:             metav1.ConditionTrue,
-			Type:               string(placementv1beta1.ApprovalRequestConditionApproved),
-			ObservedGeneration: appReq.GetGeneration(),
-			Reason:             "lgtm",
-		})
-		return hubClient.Status().Update(ctx, appReq)
-	}, updateRunEventuallyDuration, eventuallyInterval).Should(Succeed(), "Failed to get or approve approval request")
 }
